backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
gunicorn==22.0.0
websockets==15.0.1
wsproto==1.2.0
groq==0.24.0
python-dotenv==1.0.1
supabase==2.15.1
